+++ natural_language_processing/g2p_en/g2p_en_averaged_perceptron.h
@@ -21,26 +21,29 @@
 
 using namespace std;
 
+namespace ailiaG2P{
+
 class AveragedPerceptron {
 public:
-    AveragedPerceptron() {}
-    string predict(const unordered_map<string, int>& features);
+	AveragedPerceptron() {}
+	string predict(const unordered_map<string, int>& features);
 
-    unordered_map<string, unordered_map<string, double>> weights;
-    unordered_set<string> classes;
-    unordered_map<string, string> tagdict;
+	unordered_map<string, unordered_map<string, double>> weights;
+	unordered_set<string> classes;
+	unordered_map<string, string> tagdict;
 
-    void import_from_text();
+	void import_from_text();
 
-    const vector<string> START = { "-START-", "-START2-" };
-    const vector<string> END = { "-END-", "-END2-" };
+	const vector<string> START = { "-START-", "-START2-" };
+	const vector<string> END = { "-END-", "-END2-" };
 
-    unordered_map<string, int> _get_features(int i, const string& word, const vector<string>& context, const string& prev, const string& prev2);
+	unordered_map<string, int> _get_features(int i, const string& word, const vector<string>& context, const string& prev, const string& prev2);
 
-    string normalize(const string& word);
+	string normalize(const string& word);
 
-    vector<pair<string, string>> tag(const vector<string>& tokens);
+	vector<pair<string, string>> tag(const vector<string>& tokens);
 };
 
 
 int test_averaged_perceptron(void);
+}
