+++ natural_language_processing/g2p_en/g2p_en_expand.h
@@ -8,5 +8,9 @@
 *
 *******************************************************************/
 
+namespace ailiaG2P{
+
 std::string normalize_numbers(const std::string& text);
 int test_expand();
+
+}
